export_app/requirements.txt
requests
minio
psycopg2-binary
python-dotenv
pandas
